fix(review): address all PR review findings for RS-4 ephemerality contract
- Reject StaticImageNode with malformed resource_id (wrong wire length)
  instead of silently substituting all-zero hash [gemini+copilot]
- Zero out decoded_bytes on proto ingestion; field is runtime-owned,
  not to be trusted from client input [copilot]
- Change StaticImageNode.decoded_bytes to u64 to match scene budget
  accounting types and avoid lossy casts on 32-bit targets [copilot]
- Fix inaccurate doc comment on to_lowercase_hex helper [copilot]
- Rewrite no_filesystem_artifacts_created test to use isolated temp dir
  instead of walking the full workspace CWD (flaky under parallel tests) [copilot]
- Update compositor renderer tests to use new StaticImageNode API
  (resource_id + decoded_bytes, no raw blob) [copilot]
- Fix pre-existing build error: extra } closing brace in session_server.rs
  that broke mod tests scope (stranded zone/IME tests outside the module)
- Fix pre-existing test errors: missing MutationBatch.timing,
  SessionInit proto fields, StreamSession.{session_open_at_wall_us,
  dedup_window} in test struct initializations
- Fix pre-existing test error: ZoneDefinition missing ephemeral field
  in session_lifecycle.rs integration test

+++ crates/tze_hud_scene/src/graph.rs
@@ -991,7 +991,7 @@ impl SceneGraph {
         match self.nodes.get(&node_id) {
             Some(node) => {
                 let self_bytes = match &node.data {
-                    NodeData::StaticImage(img) => img.decoded_bytes as u64,
+                    NodeData::StaticImage(img) => img.decoded_bytes,
                     _ => 0,
                 };
                 self_bytes
@@ -1034,7 +1034,7 @@ impl SceneGraph {
     /// Count texture bytes in a node (not yet inserted into the graph).
     fn count_texture_bytes_in_node(node: &Node) -> u64 {
         match &node.data {
-            NodeData::StaticImage(img) => img.decoded_bytes as u64,
+            NodeData::StaticImage(img) => img.decoded_bytes,
             _ => 0,
         }
     }
@@ -3075,13 +3075,13 @@ mod tests {
     /// Per RS-4 ephemerality contract, `StaticImageNode` carries only the
     /// content-addressed `ResourceId` and the decoded byte count for budget
     /// accounting — no raw pixel data is embedded in the scene graph.
-    fn make_test_image_resource(w: u32, h: u32) -> (ResourceId, usize) {
+    fn make_test_image_resource(w: u32, h: u32) -> (ResourceId, u64) {
         // Compute a deterministic ResourceId from the dimensions (as a stand-in
         // for "the BLAKE3 hash of the actual pixel bytes").  In production this
         // would be the ResourceId returned by the resource store after upload.
         let fake_bytes: Vec<u8> = (0..w * h).flat_map(|_| [255u8, 0, 0, 255]).collect();
         let resource_id = ResourceId::of(&fake_bytes);
-        let decoded_bytes = (w * h * 4) as usize;
+        let decoded_bytes = u64::from(w * h * 4);
         (resource_id, decoded_bytes)
     }
 
@@ -3116,7 +3116,7 @@ mod tests {
             assert_eq!(si.resource_id, resource_id);
             assert_eq!(si.width, 64);
             assert_eq!(si.height, 48);
-            assert_eq!(si.decoded_bytes, 64 * 48 * 4);
+            assert_eq!(si.decoded_bytes, 64u64 * 48 * 4);
             assert_eq!(si.fit_mode, ImageFitMode::Contain);
         } else {
             panic!("expected StaticImage node data");
@@ -4533,7 +4533,7 @@ mod spec_scenarios {
                 resource_id: ResourceId::of(b"4x4 test image"),
                 width: 4,
                 height: 4,
-                decoded_bytes: 4 * 4 * 4, // 4×4 RGBA8
+                decoded_bytes: 4u64 * 4 * 4, // 4×4 RGBA8
                 fit_mode: ImageFitMode::Contain,
                 bounds: Rect::new(0.0, 0.0, 100.0, 100.0),
             }),
